+++ algorithms/src/miniweight.cpp
@@ -1,10 +1,11 @@
 #include "algorithms.hpp"
 #include <vector>
 #include <algorithm>
+#include <inttypes.h>
 
 namespace algorithms {
 
-    long long mini_weight_sum_diff(std::vector<std::pair<double, double>> nums) {
+    int64_t mini_weight_sum_diff(std::vector<std::pair<double, double>> nums) {
         auto nums_copy = nums;
 
         std::sort(nums_copy.begin(), nums_copy.end(), [](auto& a, auto& b) {
@@ -16,7 +17,7 @@ namespace algorithms {
             return da > db;
         });
 
-        long long time = 0, result = 0;
+        int64_t time = 0, result = 0;
         for (auto &[w, l] : nums_copy) {
             time += l;
             result += w * time;
@@ -25,14 +26,14 @@ namespace algorithms {
         return result;
     }
 
-    long long mini_weight_sum_scale(std::vector<std::pair<double, double>> nums) {
+    int64_t mini_weight_sum_scale(std::vector<std::pair<double, double>> nums) {
         auto nums_copy = nums;
 
         std::sort(nums_copy.begin(), nums_copy.end(), [](auto& a, auto& b) {
             return ( a.first / a.second ) < ( b.first / b.second );
         });
 
-        long long time = 0, result = 0;
+        int64_t time = 0, result = 0;
         for (auto& num : nums_copy) {
             time += num.first;
             // If you pair<int, int>, need add 1LL * time * num.second;
